transactoin amount naar decimal, plus begin validation

+++ backend/src/main/java/io/swagger/model/dto/GetTransactionDTO.java
@@ -5,6 +5,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import org.springframework.validation.annotation.Validated;
 
 import javax.validation.Valid;
+import java.math.BigDecimal;
 import java.util.Objects;
 import java.util.UUID;
 
@@ -23,7 +24,7 @@ public class GetTransactionDTO   {
   private String toIBAN = null;
 
   @JsonProperty("amount")
-  private Integer amount = null;
+  private BigDecimal amount = null;
 
   @JsonProperty("timestamp")
   private String timestamp = null;
@@ -69,7 +70,7 @@ public class GetTransactionDTO   {
     this.toIBAN = toIBAN;
   }
 
-  public GetTransactionDTO amount(Integer amount) {
+  public GetTransactionDTO amount(BigDecimal amount) {
     this.amount = amount;
     return this;
   }
@@ -80,11 +81,11 @@ public class GetTransactionDTO   {
    **/
   @Schema(description = "")
   
-    public Integer getAmount() {
+    public BigDecimal getAmount() {
     return amount;
   }
 
-  public void setAmount(Integer amount) {
+  public void setAmount(BigDecimal amount) {
     this.amount = amount;
   }
